+++ PSimulator/src/es/ull/isaatc/simulation/common/info/SynchronousInfo.java
@@ -4,7 +4,7 @@ import es.ull.isaatc.simulation.common.Simulation;
 
 public abstract class SynchronousInfo extends TimeStampedInfo {
 
-	public SynchronousInfo(Simulation simul, double ts) {
+	public SynchronousInfo(Simulation simul, long ts) {
 		super(simul, ts);
 	}
