+++ PSimulator/src/es/ull/isaatc/simulation/optThreaded/Generator.java
@@ -79,7 +79,7 @@ public abstract class Generator extends BasicElement {
          * Creates a new element-generation event.
          * @param ts Timestamp when this event must be executed.
          */
-        public GenerateEvent(double ts) {
+        public GenerateEvent(long ts) {
             super(ts, Generator.this.defLP);
         }
